tools/news-search-mcp/requirements.txt
mcp[cli]>=1.9.0,<2.0.0
httpx>=0.27.0,<1.0.0
tavily-python>=0.3.0,<1.0.0
opentelemetry-sdk>=1.24.0,<2.0.0
opentelemetry-exporter-otlp>=1.24.0,<2.0.0
langfuse>=2.0.0,<3.0.0
# Test dependencies live in requirements-test.txt — never install them in Docker images.
